restructure component folders/files (#29)
