+++ crypto-tracker/src/components/layout/Logo.jsx
