+++ crypto-tracker/src/components/shared-components/DataGridCustomNoRowsOverlay.jsx
@@ -1,7 +1,7 @@
 import { Box, CircularProgress, Typography } from "@mui/material";
 import React from "react";
-import WatchListEmptyAddButton from "./WatchListEmptyAddButton";
-import WatchListEmptyCard from "./WatchListEmptyAddButton";
+import WatchListEmptyAddButton from "../watchlist-components/WatchListEmptyAddButton";
+import WatchListEmptyCard from "../watchlist-components/WatchListEmptyAddButton";
 
 const DataGridCustomNoRowsOverlay = ({ filterText, page }) => {
   if (page === "market") {
