+++ crypto-tracker/src/components/shared-components/FilterSearchBar.jsx
