+++ crypto-tracker/src/components/watchlist-components/AddToWatchListHeaderButton.jsx
