+++ crypto-tracker/src/components/watchlist-components/AddToWatchlistChip.jsx
@@ -2,7 +2,7 @@ import { IconButton, Stack } from "@mui/material";
 import { Visibility, VisibilityOff } from "@mui/icons-material";
 import Zoom from "@mui/material/Zoom";
 import React from "react";
-import { UserContext } from "../context/UserContext";
+import { UserContext } from "../../context/UserContext";
 import { useContext } from "react";
 import { useState } from "react";
 import { useEffect } from "react";
